MOR-452: Extracted the button catalog into its own component
