+++ src/styles/_expander.scss
@@ -1,50 +1,54 @@
-.expandable {
-  cursor: pointer;
-}
+#SidePanel {
+  .expandable {
+    cursor: pointer;
+  }
 
-.expander {
-  position: absolute;
-  right: 0.8em;
-}
-.expandIcon, .collapseIcon {
-  transition: all 500ms ease;
-  opacity: 1;
-}
-:not(.expandable) > .expander {
-  display: none;
-}
+  .expander {
+    position: absolute;
+    right: 0.8em;
+  }
 
-.expandable.expanded .expandIcon,
-.expandable:not(.expanded) .collapseIcon,
-{
-  opacity: 0;
-}
+  .expandIcon, .collapseIcon {
+    transition: all 500ms ease;
+    opacity: 1;
+  }
 
-.expandable.expanded {
-  font-weight: bold;
-}
+  :not(.expandable) > .expander {
+    display: none;
+  }
 
-.expandable + .expandableContent,
-.expandable + :not(.expandableContent) > :not(.active) {
-  transition: all 500ms ease, max-height 300ms ease-out,
-  font-weight 0s, background-color 0s, color 0s;
-  opacity: 1;
-  max-height: 500px;
-}
+  .expandable.expanded .expandIcon,
+  .expandable:not(.expanded) .collapseIcon,
+  {
+    opacity: 0;
+  }
 
-.expandable:not(.expanded) + .expandableContent,
-.expandable:not(.expanded) + :not(.expandableContent) > :not(.active),
-{
-  overflow: hidden;
-  max-height: 0;
-  opacity: 0;
-  padding-top: 0;
-  padding-bottom: 0;
-  margin-top: 0;
-  margin-bottom: 0;
-}
+  .expandable.expanded {
+    font-weight: bold;
+  }
 
-.expandable.collapsed + .expandableContent,
-.expandable.collapsed + :not(.expandableContent) > :not(.active) {
-  display: none;
+  .expandable + .expandableContent,
+  .expandable + :not(.expandableContent) > :not(.active) {
+    transition: all 500ms ease, max-height 300ms ease-out,
+    font-weight 0s, background-color 0s, color 0s;
+    opacity: 1;
+    max-height: 500px;
+  }
+
+  .expandable:not(.expanded) + .expandableContent,
+  .expandable:not(.expanded) + :not(.expandableContent) > :not(.active),
+  {
+    overflow: hidden;
+    max-height: 0;
+    opacity: 0;
+    padding-top: 0;
+    padding-bottom: 0;
+    margin-top: 0;
+    margin-bottom: 0;
+  }
+
+  .expandable.collapsed + .expandableContent,
+  .expandable.collapsed + :not(.expandableContent) > :not(.active) {
+    display: none;
+  }
 }
